+++ tests/integration/it_utils.py
@@ -5,34 +5,84 @@ import inspect
 from time import sleep
 
 import xrpl  # noqa: F401 - needed for sync tests
+from xrpl.account import get_next_valid_seq_number
+from xrpl.asyncio.account import get_next_valid_seq_number as get_seq_num_async
 from xrpl.asyncio.clients import AsyncJsonRpcClient, AsyncWebsocketClient
-from xrpl.asyncio.transaction import (
-    safe_sign_and_autofill_transaction as sign_and_autofill_async,
-)
+from xrpl.asyncio.clients.async_client import AsyncClient
 from xrpl.asyncio.transaction import (
     safe_sign_and_submit_transaction as sign_and_submit_async,
 )
-from xrpl.asyncio.transaction import (
-    send_reliable_submission as send_reliable_submission_async,
-)
 from xrpl.clients import Client, JsonRpcClient, WebsocketClient
+from xrpl.models import GenericRequest, Payment
 from xrpl.models.response import Response
 from xrpl.models.transactions.transaction import Transaction
-from xrpl.transaction import (
+from xrpl.transaction import (  # noqa: F401 - needed for sync tests
     safe_sign_and_autofill_transaction,
     safe_sign_and_submit_transaction,
-    send_reliable_submission,
 )
 from xrpl.wallet import Wallet
 
-JSON_RPC_URL = "https://s.altnet.rippletest.net:51234"
-WEBSOCKET_URL = "wss://s.altnet.rippletest.net/"
+JSON_RPC_URL = "http://127.0.0.1:5005"
+WEBSOCKET_URL = "ws://127.0.0.1:6006"
+
+JSON_TESTNET_URL = "https://s.altnet.rippletest.net:51234"
+WEBSOCKET_TESTNET_URL = "wss://s.altnet.rippletest.net"
 
 JSON_RPC_CLIENT = JsonRpcClient(JSON_RPC_URL)
 ASYNC_JSON_RPC_CLIENT = AsyncJsonRpcClient(JSON_RPC_URL)
 
 WEBSOCKET_CLIENT = WebsocketClient(WEBSOCKET_URL)
 ASYNC_WEBSOCKET_CLIENT = AsyncWebsocketClient(WEBSOCKET_URL)
+
+JSON_RPC_TESTNET_CLIENT = JsonRpcClient(JSON_TESTNET_URL)
+ASYNC_JSON_RPC_TESTNET_CLIENT = AsyncJsonRpcClient(JSON_TESTNET_URL)
+
+WEBSOCKET_TESTNET_CLIENT = WebsocketClient(WEBSOCKET_TESTNET_URL)
+ASYNC_WEBSOCKET_TESTNET_CLIENT = AsyncWebsocketClient(WEBSOCKET_TESTNET_URL)
+
+# (is_async, is_json, is_testnet) -> client
+_CLIENTS = {
+    (True, True, True): ASYNC_JSON_RPC_TESTNET_CLIENT,
+    (True, True, False): ASYNC_JSON_RPC_CLIENT,
+    (True, False, True): ASYNC_WEBSOCKET_TESTNET_CLIENT,
+    (True, False, False): ASYNC_WEBSOCKET_CLIENT,
+    (False, True, True): JSON_RPC_TESTNET_CLIENT,
+    (False, True, False): JSON_RPC_CLIENT,
+    (False, False, True): WEBSOCKET_TESTNET_CLIENT,
+    (False, False, False): WEBSOCKET_CLIENT,
+}
+
+MASTER_ACCOUNT = "rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh"
+MASTER_SECRET = "snoPBrXtMeMyMHUVTgbuqAfg1SUTb"
+MASTER_WALLET = Wallet(MASTER_SECRET, 0)
+FUNDING_AMOUNT = "1200000000"
+
+LEDGER_ACCEPT_REQUEST = GenericRequest(method="ledger_accept")
+
+
+def fund_wallet_sync(wallet: Wallet) -> None:
+    client = JSON_RPC_CLIENT
+    payment = Payment(
+        account=MASTER_ACCOUNT,
+        destination=wallet.classic_address,
+        amount=FUNDING_AMOUNT,
+    )
+    safe_sign_and_submit_transaction(payment, MASTER_WALLET, client, check_fee=True)
+    client.request(LEDGER_ACCEPT_REQUEST)
+    wallet.sequence = get_next_valid_seq_number(wallet.classic_address, client)
+
+
+async def fund_wallet(
+    wallet: Wallet, client: AsyncClient = ASYNC_JSON_RPC_CLIENT
+) -> None:
+    payment = Payment(
+        account=MASTER_ACCOUNT,
+        destination=wallet.classic_address,
+        amount=FUNDING_AMOUNT,
+    )
+    await sign_and_submit_async(payment, MASTER_WALLET, client, check_fee=True)
+    await client.request(LEDGER_ACCEPT_REQUEST)
+    wallet.sequence = await get_seq_num_async(wallet.classic_address, client)
 
 
 def submit_transaction(
@@ -60,34 +110,39 @@ def sign_and_reliable_submission(
     transaction: Transaction, wallet: Wallet, use_json_client: bool = True
 ) -> Response:
     client = _choose_client(use_json_client)
-    signed_tx = safe_sign_and_autofill_transaction(transaction, wallet, client)
-    return send_reliable_submission(signed_tx, client)
+    response = submit_transaction(transaction, wallet, client)
+    client.request(LEDGER_ACCEPT_REQUEST)
+    return response
 
 
 async def sign_and_reliable_submission_async(
     transaction: Transaction, wallet: Wallet, use_json_client: bool = True
 ) -> Response:
     client = _choose_client_async(use_json_client)
-    signed_tx = await sign_and_autofill_async(transaction, wallet, client)
-    return await send_reliable_submission_async(signed_tx, client)
+    response = await submit_transaction_async(transaction, wallet, client)
+    await client.request(LEDGER_ACCEPT_REQUEST)
+    return response
 
 
 def _choose_client(use_json_client: bool) -> Client:
-    if use_json_client:
-        return JSON_RPC_CLIENT
-    else:
-        return WEBSOCKET_CLIENT
+    return _CLIENTS[(False, use_json_client, False)]
 
 
 def _choose_client_async(use_json_client: bool) -> Client:
-    if use_json_client:
-        return ASYNC_JSON_RPC_CLIENT
-    else:
-        return ASYNC_WEBSOCKET_CLIENT
+    return _CLIENTS[(True, use_json_client, False)]
 
 
+def _get_client(is_async: bool, is_json: bool, is_testnet: bool) -> Client:
+    return _CLIENTS[(is_async, is_json, is_testnet)]
+
+
+# TODO: document how to write tests, for posterity
 def test_async_and_sync(
-    original_globals, modules=None, websockets_only=False, num_retries=1
+    original_globals,
+    modules=None,
+    websockets_only=False,
+    num_retries=1,
+    use_testnet=False,
 ):
     def decorator(test_function):
         lines = _get_non_decorator_code(test_function)
@@ -154,14 +209,18 @@ def test_async_and_sync(
 
         def modified_test(self):
             if not websockets_only:
-                with self.subTest(version="sync", client="json"):
-                    _run_sync_test(self, JSON_RPC_CLIENT)
                 with self.subTest(version="async", client="json"):
-                    asyncio.run(_run_async_test(self, ASYNC_JSON_RPC_CLIENT))
-            with self.subTest(version="sync", client="websocket"):
-                _run_sync_test(self, WEBSOCKET_CLIENT)
+                    asyncio.run(
+                        _run_async_test(self, _get_client(True, True, use_testnet))
+                    )
+                with self.subTest(version="sync", client="json"):
+                    _run_sync_test(self, _get_client(False, True, use_testnet))
             with self.subTest(version="async", client="websocket"):
-                asyncio.run(_run_async_test(self, ASYNC_WEBSOCKET_CLIENT))
+                asyncio.run(
+                    _run_async_test(self, _get_client(True, False, use_testnet))
+                )
+            with self.subTest(version="sync", client="websocket"):
+                _run_sync_test(self, _get_client(False, False, use_testnet))
 
         return modified_test