+++ tests/integration/reusable_values.py
@@ -1,17 +1,27 @@
 import asyncio
 
 from tests.integration.it_utils import (
-    ASYNC_JSON_RPC_CLIENT,
+    ASYNC_JSON_RPC_TESTNET_CLIENT,
+    fund_wallet,
     sign_and_reliable_submission_async,
 )
 from xrpl.asyncio.wallet import generate_faucet_wallet
 from xrpl.models.amounts import IssuedCurrencyAmount
 from xrpl.models.transactions import OfferCreate, PaymentChannelCreate
+from xrpl.wallet import Wallet
 
 
+# TODO: use `asyncio.gather` for these, to parallelize
+# TODO: set up wallet for each test instead of using one for all tests (now that it's
+# faster)
 async def _set_up_reusable_values():
-    WALLET = await generate_faucet_wallet(ASYNC_JSON_RPC_CLIENT)
-    DESTINATION = await generate_faucet_wallet(ASYNC_JSON_RPC_CLIENT)
+    WALLET = Wallet.create()
+    await fund_wallet(WALLET)
+    DESTINATION = Wallet.create()
+    await fund_wallet(DESTINATION)
+
+    TESTNET_WALLET = await generate_faucet_wallet(ASYNC_JSON_RPC_TESTNET_CLIENT)
+    TESTNET_DESTINATION = await generate_faucet_wallet(ASYNC_JSON_RPC_TESTNET_CLIENT)
 
     OFFER = await sign_and_reliable_submission_async(
         OfferCreate(
@@ -41,12 +51,21 @@ async def _set_up_reusable_values():
     )
     WALLET.sequence += 1
 
-    return WALLET, DESTINATION, OFFER, PAYMENT_CHANNEL
+    return (
+        WALLET,
+        DESTINATION,
+        TESTNET_WALLET,
+        TESTNET_DESTINATION,
+        OFFER,
+        PAYMENT_CHANNEL,
+    )
 
 
 (
     WALLET,
     DESTINATION,
+    TESTNET_WALLET,
+    TESTNET_DESTINATION,
     OFFER,
     PAYMENT_CHANNEL,
 ) = asyncio.run(_set_up_reusable_values())