+++ xrpl/asyncio/clients/websocket_base.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import json
 from asyncio import Future, Queue, Task, create_task, get_running_loop
 from random import randrange
-from typing import Any, Dict, Optional
+from typing import TYPE_CHECKING, Any, Dict, Optional, cast
 
 from typing_extensions import Final
 from websockets.legacy.client import WebSocketClientProtocol, connect
@@ -16,6 +16,16 @@ from xrpl.models.requests.request import Request
 from xrpl.models.response import Response
 
 _REQ_ID_MAX: Final[int] = 1_000_000
+# the types from asyncio are not implemented as generics in python 3.8 and
+# lower, so we need to only subscript them when running typechecking.
+if TYPE_CHECKING:
+    _REQUESTS_TYPE = Dict[str, Future[Dict[str, Any]]]
+    _MESSAGES_TYPE = Queue[Dict[str, Any]]
+    _HANDLER_TYPE = Task[None]
+else:
+    _REQUESTS_TYPE = Dict[str, Future]
+    _MESSAGES_TYPE = Queue
+    _HANDLER_TYPE = Task
 
 
 def _inject_request_id(request: Request) -> Request:
@@ -28,9 +38,7 @@ def _inject_request_id(request: Request) -> Request:
         return request
     request_dict = request.to_dict()
     request_dict["id"] = f"{request.method}_{randrange(_REQ_ID_MAX)}"
-    resp = Request.from_dict(request_dict)
-    assert resp.id is not None  # mypy
-    return resp
+    return Request.from_dict(request_dict)
 
 
 class WebsocketBase(Client):
@@ -47,10 +55,14 @@ class WebsocketBase(Client):
         Arguments:
             url: The URL of the rippled node to submit requests to.
         """
-        self._open_requests: Dict[str, Future[Dict[str, Any]]] = {}
+        self._open_requests: _REQUESTS_TYPE = {}
         self._websocket: Optional[WebSocketClientProtocol] = None
-        self._handler_task: Optional[Task[None]] = None
-        self._messages: Optional[Queue[Dict[str, Any]]] = None
+        self._handler_task: Optional[_HANDLER_TYPE] = None
+        # unfortunately, we cannot create the Queue here because it needs to be
+        # tied to a currently-running event loop. the sync websocket client
+        # will initialize a new event loop when it opens the connection, so for
+        # that client the initializer cannot create the queue
+        self._messages: Optional[_MESSAGES_TYPE] = None
         super().__init__(url)
 
     def is_open(self: WebsocketBase) -> bool:
@@ -69,9 +81,6 @@ class WebsocketBase(Client):
 
     async def _do_open(self: WebsocketBase) -> None:
         """Connects the client to the Web Socket API at its URL."""
-        if self.is_open():
-            return
-
         # open the connection
         self._websocket = await connect(self.url)
 
@@ -83,14 +92,8 @@ class WebsocketBase(Client):
 
     async def _do_close(self: WebsocketBase) -> None:
         """Closes the connection."""
-        if not self.is_open():
-            return
-        assert self._handler_task is not None  # mypy
-        assert self._websocket is not None  # mypy
-        assert self._messages is not None  # mypy
-
         # cancel the handler
-        self._handler_task.cancel()
+        cast(_HANDLER_TYPE, self._handler_task).cancel()
         self._handler_task = None
 
         # cancel any pending request Futures
@@ -99,13 +102,13 @@ class WebsocketBase(Client):
         self._open_requests = {}
 
         # clear the message queue
-        for _ in range(self._messages.qsize()):
-            self._messages.get_nowait()
-            self._messages.task_done()
+        for _ in range(cast(_MESSAGES_TYPE, self._messages).qsize()):
+            cast(_MESSAGES_TYPE, self._messages).get_nowait()
+            cast(_MESSAGES_TYPE, self._messages).task_done()
         self._messages = None
 
         # close the connection
-        await self._websocket.close()
+        await cast(WebSocketClientProtocol, self._websocket).close()
 
     async def _handler(self: WebsocketBase) -> None:
         """
@@ -117,9 +120,7 @@ class WebsocketBase(Client):
 
         As long as a given client remains open, this handler will be running as a Task.
         """
-        assert self._websocket is not None  # mypy
-        assert self._messages is not None  # mypy
-        async for response in self._websocket:
+        async for response in cast(WebSocketClientProtocol, self._websocket):
             response_dict = json.loads(response)
 
             # if this response corresponds to request, fulfill the Future
@@ -127,7 +128,7 @@ class WebsocketBase(Client):
                 self._open_requests[response_dict["id"]].set_result(response_dict)
 
             # enqueue the response for the message queue
-            self._messages.put_nowait(response_dict)
+            cast(_MESSAGES_TYPE, self._messages).put_nowait(response_dict)
 
     def _set_up_future(self: WebsocketBase, request: Request) -> None:
         """
@@ -148,8 +149,7 @@ class WebsocketBase(Client):
         self._open_requests[request_str] = get_running_loop().create_future()
 
     async def _do_send_no_future(self: WebsocketBase, request: Request) -> None:
-        assert self._websocket is not None  # mypy
-        await self._websocket.send(
+        await cast(WebSocketClientProtocol, self._websocket).send(
             json.dumps(
                 request_to_websocket(request),
             ),
@@ -162,14 +162,13 @@ class WebsocketBase(Client):
         await self._do_send_no_future(request)
 
     async def _do_pop_message(self: WebsocketBase) -> Dict[str, Any]:
-        assert self._messages is not None  # mypy
-        msg = await self._messages.get()
-        self._messages.task_done()
+        msg = await cast(_MESSAGES_TYPE, self._messages).get()
+        cast(_MESSAGES_TYPE, self._messages).task_done()
         return msg
 
-    async def request_impl(self: WebsocketBase, request: Request) -> Response:
+    async def _do_request_impl(self: WebsocketBase, request: Request) -> Response:
         """
-        Base ``request_impl`` implementation for Websockets.
+        Base ``request_impl`` implementation for websockets.
 
         Arguments:
             request: An object representing information about a rippled request.
@@ -180,12 +179,7 @@ class WebsocketBase(Client):
         Raises:
             XRPLWebsocketException: If there is already an open request by the
                 request's ID, or if this WebsocketBase is not open.
-
-        :meta private:
         """
-        if not self.is_open():
-            raise XRPLWebsocketException("Websocket is not open")
-
         # if no ID on this request, generate and inject one, and ensure it
         # is backed by a future
         request_with_id = _inject_request_id(request)