+++ xrpl/models/transactions/__init__.py
@@ -23,7 +23,8 @@ from xrpl.models.transactions.payment_channel_create import PaymentChannelCreate
 from xrpl.models.transactions.payment_channel_fund import PaymentChannelFund
 from xrpl.models.transactions.set_regular_key import SetRegularKey
 from xrpl.models.transactions.signer_list_set import SignerEntry, SignerListSet
-from xrpl.models.transactions.transaction import Memo, Signer
+from xrpl.models.transactions.ticket_create import TicketCreate
+from xrpl.models.transactions.transaction import Memo, Signer, Transaction
 from xrpl.models.transactions.trust_set import TrustSet, TrustSetFlag
 
 __all__ = [
@@ -51,6 +52,8 @@ __all__ = [
     "Signer",
     "SignerEntry",
     "SignerListSet",
+    "TicketCreate",
+    "Transaction",
     "TrustSet",
     "TrustSetFlag",
 ]