+++ CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [[Unreleased]]
+### Added
+- `xrpl.utils.str_to_hex` and `xrpl.utils.hex_to_str` helpers
+- `ledger_index` optional param for all the main account methods
+- `TicketCreate` transaction model
+- `GenericRequest` model for unsupported request types
+- Methods to convert between `IssuedCurrency` and `IssuedCurrencyAmount`
+- Support for ints and floats in the `IssuedCurrency` and `IssuedCurrencyAmount` models (and ints for `XRP`)
+
+### Fixed
+- Makes the default ledger version for `get_next_valid_seq_number` `current` instead of `validated`
+- Stops erroring on non-`tesSUCCESS` responses in reliable transaction submission
+- Removes runtime asserts in websocket clients that were used for type checks
+  only
+- Adds missing top-level `py.typed` file for exceptions and constants
+- Fix issue where unsupported currency codes weren't being correctly processed in the binary codec
+- Fixes issue with UNLModify encoding (due to a bug in rippled)
+- Exports `Transaction`, `Response`, pseudo-transactions at the `xrpl.models` level
 
 ## [1.2.0] - 2021-11-09
 ### Added